hospital/requirements.txt
BeautifulSoup==3.2.1
colorama==0.3.7
Django==1.8
django-tastypie==0.13.3
Jinja2==2.8
MarkupSafe==0.23
nltk==3.2.1
paho-mqtt==1.1
Pillow==3.1.1
psycopg2==2.6.1
python-dateutil==2.5.1
python-mimeparse==1.5.1
reportlab==3.3.0
requests==2.9.1
selenium==2.53.6
six==1.10.0
termcolor==1.1.0
pycrypto==2.6.1
docopt==0.6.2
elasticsearch==1.2.0
lxml==3.6.0
pytest
python-owasp-zap-v2.4==0.0.8


